574-Assignment/pyproject.toml
[project]
name = "574-assignment-rag-comparison"
version = "0.1.0"
description = "D&D 5e RAG System Comparison Framework"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pymilvus[milvus_lite]>=2.4.0",
    "sentence-transformers>=2.2.2",
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "scipy>=1.11.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = []  # Not a package, just a script collection

